// front/src/enums/StoragePathEnum.ts
export enum StoragePathEnum {
  configuration = '/configuration',

  files = '/files',
  files_features = '/files_features',
  files_timestamps = '/files_timestamps',
  files_sites = '/files_sites',
  files_tags = '/files_tags',
  files_metas = '/files_metas',

  meta_properties = '/meta_properties',
  meta_sets = '/meta_sets',

  ranges = '/ranges',
  ranges_timestamps = '/ranges_timestamps',

  bands = '/bands',
  bands_frequencies = '/bands_frequencies',

  umaps = '/umaps',
  umaps_integrations = '/umaps_integrations',
  umaps_bands = '/umaps_bands',
  umaps_ranges = '/umaps_ranges',
  umaps_sites = '/umaps_sites',

  groups_features = '/groups_features',
  groups_timestamps = '/groups_timestamps',
  groups_features_reduced = '/groups_features_reduced',
}